Improvements around Doctrine
Add contract for Unit of Work
Add package for integration tests
Enable databse integration tests in CI
Fix folder structure
Enable phpstan doctrine extension
