+++ webservice/config/packages/doctrine.yaml
@@ -8,9 +8,9 @@ doctrine:
 
         profiling_collect_backtrace: '%kernel.debug%'
         types:
-            instant:  'App\Infrastructure\Doctrine\DBAL\Type\InstantType'
-            local_datetime:  'App\Infrastructure\Doctrine\DBAL\Type\LocalDateTimeType'
-            local_time:  'App\Infrastructure\Doctrine\DBAL\Type\LocalTimeType'
+            instant:  'App\Infrastructure\Persistence\DoctrineORM\Type\InstantType'
+            local_datetime:  'App\Infrastructure\Persistence\DoctrineORM\Type\LocalDateTimeType'
+            local_time:  'App\Infrastructure\Persistence\DoctrineORM\Type\LocalTimeType'
         use_savepoints: true
     orm:
         auto_generate_proxy_classes: true
