+++ webservice/config/packages/zenstruck_foundry.yaml
@@ -1,0 +1,5 @@
+when@dev: &dev
+    # See full configuration: https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#full-default-bundle-configuration
+    zenstruck_foundry:
+
+when@test: *dev
