+++ webservice/src/Core/Contracts/Persistence/UnitOfWork.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Core\Contracts\Persistence;
+
+interface UnitOfWork
+{
+    public function commit(): void;
+}
